Float Max Min

+++ MinMax.cs
@@ -8,7 +8,7 @@ namespace MinMaxProblems
 {
     public class MinMax
     {
-        public static int MaximumIntNum(int fNumber, int sNumber, int thirdNumber)
+        public static double MaximumIntNum(double fNumber, double sNumber, double thirdNumber)
         {
             if ((fNumber.CompareTo(sNumber) > 0 && fNumber.CompareTo(thirdNumber) > 0) ||
                 (fNumber.CompareTo(sNumber) >= 0 && fNumber.CompareTo(thirdNumber) > 0) ||
